Rename directory from action to actions
