+++ javascriptv3/example_code/bedrock-runtime/actions/invoke-claude.js
