+++ javascriptv3/example_code/bedrock-runtime/actions/invoke-jurassic2.js
